Extract general functionality into BaseTabActivity
