+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/stats/SoldierStatisticsActivity.java
@@ -1,91 +1,33 @@
 package com.ninetwozero.bf4intel.ui.stats;
 
-import android.app.ActionBar;
-import android.app.FragmentTransaction;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.base.ui.BaseIntelActivity;
+import com.ninetwozero.bf4intel.base.ui.BaseTabActivity;
 import com.ninetwozero.bf4intel.factories.FragmentFactory;
-import com.ninetwozero.bf4intel.ui.adapters.ViewPagerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class SoldierStatisticsActivity extends BaseIntelActivity implements ActionBar.TabListener {
-    public static final String INTENT_PROFILE = "profile";
+public class SoldierStatisticsActivity extends BaseTabActivity {
     private static final int[] TITLES = new int[]{
         R.string.weapons,
         R.string.vehicles,
         R.string.reports
     };
-    private ViewPager viewPager;
-    private ViewPagerAdapter viewPagerAdapter;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.generic_multi_tab_activity);
-        initialize();
+    protected int[] getTitleResources() {
+        return TITLES;
     }
 
     @Override
-    public void onTabSelected(final ActionBar.Tab tab, final FragmentTransaction fragmentTransaction) {
-        viewPager.setCurrentItem(tab.getPosition(), true);
-    }
-
-    @Override
-    public void onTabUnselected(final ActionBar.Tab tab, final FragmentTransaction fragmentTransaction) {
-    }
-
-    @Override
-    public void onTabReselected(final ActionBar.Tab tab, final FragmentTransaction fragmentTransaction) {
-    }
-
-    private void initialize() {
-        setupViewPagerAdapter();
-        setupViewPager();
-        setupActionBar();
-    }
-
-    private void setupViewPagerAdapter() {
-        viewPagerAdapter = new ViewPagerAdapter(getSupportFragmentManager(), generateFragmentList());
-    }
-
-    private List<Fragment> generateFragmentList() {
+    protected List<Fragment> fetchFragmentsForActivity(final Bundle profileBundle) {
         final List<Fragment> fragments = new ArrayList<Fragment>();
-        final Bundle profileBundle = getIntent().getBundleExtra(INTENT_PROFILE);
-
         fragments.add(FragmentFactory.get(FragmentFactory.Type.WEAPON_STATS, profileBundle));
         fragments.add(FragmentFactory.get(FragmentFactory.Type.VEHICLE_STATS, profileBundle));
         fragments.add(FragmentFactory.get(FragmentFactory.Type.SOLDIER_STATS, profileBundle));
-
         return fragments;
-    }
-
-    private void setupViewPager() {
-        viewPager = (ViewPager) findViewById(R.id.viewpager);
-        viewPager.setOffscreenPageLimit(viewPagerAdapter.getCount());
-        viewPager.setOnPageChangeListener(
-                new ViewPager.SimpleOnPageChangeListener() {
-                    @Override
-                    public void onPageSelected(int position) {
-                        getActionBar().setSelectedNavigationItem(position);
-                    }
-                }
-        );
-        viewPager.setAdapter(viewPagerAdapter);
-    }
-
-    private void setupActionBar() {
-        final ActionBar actionBar = getActionBar();
-        actionBar.setDisplayShowHomeEnabled(true);
-        actionBar.setDisplayHomeAsUpEnabled(true);
-        actionBar.setNavigationMode(ActionBar.NAVIGATION_MODE_TABS);
-        for (int i = 0, max = TITLES.length; i < max; i++) {
-            actionBar.addTab(actionBar.newTab().setText(TITLES[i]).setTabListener(this));
-        }
     }
 }
